feat(testes): update tests
